requirements.txt
asgiref==3.5.2
crispy-bootstrap5==0.6
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==1.0.0
Django==4.1
django-appconf==1.0.5
django-autocomplete-light==3.9.4
django-crispy-forms==1.14.0
django-import-export==2.9.0
django-select2-forms==3.0.0
django-sortedm2m==3.1.1
et-xmlfile==1.1.0
gunicorn==20.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.10
psycopg2==2.9.3
psycopg2-binary==2.9.3
pyparsing==3.0.9
python-dateutil==2.8.2
PyYAML==6.0
six==1.16.0
sqlparse==0.4.2
tablib==3.2.1
tzdata==2022.2
whitenoise==6.2.0
xlrd==2.0.1
xlwt==1.3.0
